Update the flexmeson windows binary version
The old binary/url is no longer available, and the layout of the
sourceforge archvie seems to have changed to include a new
old_versions/ directory where the previous binary is
relocated.

https://sourceforge.net/projects/winflexbison/files/old_versions/

While we could use that, it seem better to invest the effort to
adapt the scipt afterwards to use the github archives instead

https://github.com/lexxmark/winflexbison/releases

So for now bump the version to get the subproject building
till we switch to github releases.